# app/static/swagger.yaml
...
swagger: "2.0"
...
info:
  title: "Mechanic API" # Title of the API
  description: "API to manage a Mechanic Shop" # A brief description of what the API does
  version: "1.0.0" # The version of your API

# The base URL of the API, in this case, "127.0.0.1:5000". This is where the API can be accessed.
host: "127.0.0.1:5000"
...
schemes:
  - "http"

# The types of request body formats the API can consume. Here, it's "application/json",
# meaning the API can accept JSON-formatted request bodies.
consumes:
  - "application/json"

# The types of response body formats the API can produce. Here, it's also "application/json",
# indicating that the API will respond with JSON-formatted data.
produces:
  - "application/json"

# Add securityDefinitions to allow endpoints to require jwt (i.e. tokens) for access
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header

# Paths - This is where you would define the endpoints of your API, along with the HTTP methods
# they support and their respective request/response formats.

# -------------------- Customer Endpoints --------------------
paths:
...
  /mechanics/login: # ⚡ Tested!
    post:
      tags:
        - Mechanics
      summary: "Endpoint to login a mechanic (generates token)"
      description: "Endpoint to login a mechanic in our Mechanic Shop system. Token is required."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new mechanic"
          required: true
          schema:
            type: "object"
            properties:
              email:
                type: "string"
              password:
                type: "string"
      responses:
        200:
          description: "Successful login"
          schema:
            type: "object"
            properties:
              token:
                type: "string"
              message:
                type: "string"
              status:
                type: "string"
          examples:
            application/json:
              token: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoxLCJleHAiOjE2ODgwMDAwMDB9.XYZ123abc456DEF789ghi012JKL345mno678PQR90stu"
              message: "Login successful"
              status: "success"
  /mechanics: # ⚡ Tested!
    post:
      tags:
        - Mechanics
      summary: "Endpoint to create/register a new mechanic"
      description: "Endpoint to create a new mechanic in our Mechanic Shop system. Emails must be unique."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new mechanic"
          required: true
          schema:
            $ref: "#/definitions/MechanicPayload"
      responses:
        201:
          description: "Mechanic created successfully"
          schema:
            $ref: "#/definitions/MechanicResponse"
          examples:
            application/json:
              id: 1
              name: "Mechanic 1"
              email: "mechanic1@email.com"
              phone: "123-456-7890"
              password: "securepassword123"
              salary: 60000
              service_tickets: [1, 2, 3]

    get: # ⚡ Tested!
      tags:
        - Mechanics
      summary: Get mechanics
      description: Retrieve a list of all mechanics in the system.
      responses:
        200:
          description: "Retrieved Mechanics Successfully"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/MechanicListResponse"

    put: # ⚡ Tested!
      tags:
        - Mechanics
      summary: "Endpoint to update a mechanic."
      description: "Endpoint to update a mechanic. This is a token authenticated route, and the mechanic must be logged in to update their account"
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to update a mechanic."
          required: true
          schema:
            $ref: "#/definitions/UpdateMechanicPayload"
      responses:
        200:
          description: "Mechanic updated successfully"
          schema:
            $ref: "#/definitions/UpdateMechanicResponse"
          examples:
            application/json:
              id: 1
              name: "Mechanic 1"
              email: "mechanic1@email.com"
              phone: "123-456-7890"
              password: "securepassword123"
              salary: 60000
              service_tickets: [1, 2, 3]

    delete: # ⚡ Tested!
      tags:
        - Mechanics
      summary: "Endpoint to delete a mechanic."
      description: "Endpoint to delete a mechanic. This is a token authenticated route, and the mechanic must be logged in to delete their account."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Mechanic deleted successfully"
          schema:
            $ref: "#/definitions/DeleteMechanicResponse"
          examples:
            application/json:
              message: "Mechanic <mechanic_id> deleted successfully."
              status: "success"
  /mechanics/{mechanic_id}: # ⚡ Tested!
    get:
      tags:
        - Mechanics
      summary: "Get a specific mechanic by ID"
      description: "Retrieve a specific mechanic using their unique ID."
      parameters:
        - name: mechanic_id
          in: path
          required: true
          description: "ID of mechanic to retrieve"
          type: integer
      responses:
        200:
          description: "Mechanic retrieved successfully"
          schema:
            $ref: "#/definitions/MechanicResponse"
          examples:
            application/json:
              id: 1
              name: "Mechanic 1"
              email: "mechanic1@email.com"
              phone: "123-456-7890"
              password: "securepassword123"
              salary: 60000
              service_tickets: [1, 2, 3]
        404:
          description: "Mechanic not found"
  /mechanics/{mechanic_id}/add-ticket/{service_ticket_id}: # ⚡ Tested!
    post:
      tags:
        - Mechanics
      summary: "Endpoint to add a mechanic to a service ticket"
      description: "Endpoint to add a mechanic to a service ticket in our Mechanic Shop system. Token is required."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: mechanic_id
          required: true
          description: "Enter the mechanic ID you want to assign to the service ticket"
          type: integer

        - in: path
          name: service_ticket_id
          required: true
          description: "Enter the service ticket ID you want to assign the mechanic to"
          type: integer
      responses:
        201:
          description: "Mechanic added successfully"
          schema:
            $ref: "#/definitions/MechanicResponse"
          examples:
            application/json:
              message: "Mechanic <mechanic_id> assigned to ticket <service_ticket_id>."
              status: "success"
  /mechanics/{mechanic_id}/remove-ticket/{service_ticket_id}: # ⚡ Tested!
    delete:
      tags:
        - Mechanics
      summary: "Endpoint to remove a mechanic from a service ticket"
      description: "Endpoint to remove a mechanic from a service ticket in our Mechanic Shop system. Token is required."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: mechanic_id
          required: true
          description: "Enter the mechanic ID you want to remove from the service ticket"
          type: integer

        - in: path
          name: service_ticket_id
          required: true
          description: "Enter the service ticket ID you want to remove the mechanic from"
          type: integer
      responses:
        201:
          description: "Mechanic removed successfully"
          schema:
            $ref: "#/definitions/MechanicResponse"
          examples:
            application/json:
              message: "Mechanic <mechanic_id> removed from ticket <service_ticket_id>."
              status: "success"
...
definitions:
...
  MechanicPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
      salary:
        type: "integer"
  MechanicResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
      salary:
        type: "integer"
      service_tickets:
        type: "array"
        items:
          type: "integer"
  MechanicListResponse:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        email:
          type: "string"
        phone:
          type: "string"
        salary:
          type: "integer"
        password:
          type: "string"
        service_tickets:
          type: "array"
          items:
            type: "integer"
  UpdateMechanicPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
      salary:
        type: "integer"
    required:
      - name
      - email
      - phone
      - password
      - salary
  UpdateMechanicResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
      salary:
        type: "integer"
  DeleteMechanicResponse:
    type: "object"
    properties:
      message:
        type: "string"
      status:
        type: "string"
  # -------------------- Service Ticket Definitions --------------------
...